# services/payments-svc/internal/server/openapi.yaml
...
openapi: 3.0.3
info:
    title: Omnibus network payment hub API
    version: 1.0.0
paths:
...
    /v1/operator/clock:
        post:
            tags:
                - NetworkOperationsService
            description: |-
                SetClock moves the scenario clock (and the chain's) forward. Demo builds
                 only; a production build answers PERMISSION_DENIED.
            operationId: NetworkOperationsService_SetClock
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.SetClockRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.SetClockResponse'
    /v1/operator/netting-cycles:
        post:
            tags:
                - NetworkOperationsService
            description: RunNettingCycle plans and settles a netting cycle now.
            operationId: NetworkOperationsService_RunNettingCycle
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.RunNettingCycleRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.NettingCycle'
...
    /v1/operator/reconciliations:
        post:
            tags:
                - NetworkOperationsService
            description: Reconcile attests the Fed's camt.052 balance for the joint account.
            operationId: NetworkOperationsService_Reconcile
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.payments.v1.ReconcileRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.payments.v1.Reconciliation'
...
components:
    schemas:
...
        api.payments.v1.Money:
            type: object
            properties:
                amount:
                    type: string
                currency:
                    type: string
            description: |-
                Money is an amount in a currency. The amount is a decimal string with at
                 most two fraction digits ("25000000.00"), never a float.
        api.payments.v1.NettingCycle:
            type: object
            properties:
                id:
                    type: string
                cycleRef:
                    type: string
                at:
                    type: string
                    format: date-time
                trigger:
                    type: string
                discharged:
                    type: integer
                    format: int32
                deferred:
                    type: integer
                    format: int32
                gross:
                    $ref: '#/components/schemas/api.payments.v1.Money'
                net:
                    $ref: '#/components/schemas/api.payments.v1.Money'
...
        api.payments.v1.ReconcileRequest:
            type: object
            properties: {}
        api.payments.v1.Reconciliation:
            type: object
            properties:
                id:
                    type: string
                at:
                    type: string
                    format: date-time
                fedBalance:
                    $ref: '#/components/schemas/api.payments.v1.Money'
                ledgerTotal:
                    $ref: '#/components/schemas/api.payments.v1.Money'
                reconciliationBreak:
                    type: boolean
                invariantsHold:
                    type: boolean
...
        api.payments.v1.RunNettingCycleRequest:
            type: object
            properties: {}
        api.payments.v1.SetClockRequest:
            type: object
            properties:
                time:
                    type: string
                    format: date-time
        api.payments.v1.SetClockResponse:
            type: object
            properties:
                now:
                    type: string
                    format: date-time
...